requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
WTForms==3.1.0
qrcode==7.4.2
Pillow==10.1.0
python-escpos==3.0a9
pycryptodome==3.19.0
Werkzeug==3.0.1
APScheduler==3.10.4
python-dotenv==1.0.0
openpyxl==3.1.2
firebase-admin==6.5.0
Authlib==1.3.2
gunicorn==22.0.0

